Approve action modals in admin panel and personal api token settings (#59)
* Disable hash in version
* Change new page default title and content
* Admin: Users and Groups action now have approve modal
* Profile: revoke token now have approve modal
* Fix favorites ordering ( order changed between requests)
* Fix users ordering to
* Remove markdown in favorite page card (in favorites ui)

+++ src/Mimisbrunnr.Web.Host/ClientApp/src/components/people/profile/settings/ApiTokens.vue
@@ -60,6 +60,18 @@ export default {
       this.tokens = tokenRequest.data;
     },
     revoke: async function (token) {
+      var approve = await this.$bvModal.msgBoxConfirm(this.$t("profile.settings.tokens.approveModal.revoke"), {
+            title: this.$t("profile.settings.tokens.approveModal.title"),
+            centered: true,
+            size: 'sm',
+            buttonSize: 'sm',
+            cancelTitle: this.$t("profile.settings.tokens.approveModal.cancel"),
+            okTitle: this.$t("profile.settings.tokens.approveModal.ok"),
+            okVariant: 'danger',
+            headerClass: 'p-2 border-bottom-0',
+            footerClass: 'p-2 border-top-0',
+      });
+      if(!approve) return;
       await axios.delete(`/api/account/token/${token.id}`);
       await this.load();
     },
